[SW-1787] Obtaining nodes can be done via rest API in all cases on External backend
