+++ core/src/main/scala/org/apache/spark/h2o/backends/external/ExternalBackendConf.scala
@@ -61,7 +61,7 @@ trait ExternalBackendConf extends SharedBackendConf {
   def externalCommunicationBlockSize: String = sparkConf.get(PROP_EXTERNAL_COMMUNICATION_BLOCK_SIZE._1, PROP_EXTERNAL_COMMUNICATION_BLOCK_SIZE._2)
   def externalBackendStopTimeout: Int = sparkConf.getInt(PROP_EXTERNAL_BACKEND_STOP_TIMEOUT._1, PROP_EXTERNAL_BACKEND_STOP_TIMEOUT._2)
   private[backends] def isBackendVersionCheckDisabled() = sparkConf.getBoolean(PROP_EXTERNAL_DISABLE_VERSION_CHECK._1, PROP_EXTERNAL_DISABLE_VERSION_CHECK._2)
-  private[backends] def externalBackendFlatFileName() = sparkConf.getOption(PROP_EXTERNAL_FLATFILE_NAME._1)
+
   /** Setters */
 
   /**
@@ -119,7 +119,6 @@ trait ExternalBackendConf extends SharedBackendConf {
   def setExternalExtraMemoryPercent(memoryPercent: Int): H2OConf = set(PROP_EXTERNAL_EXTRA_MEMORY_PERCENT._1, memoryPercent.toString)
   def setExternalCommunicationBlockSize(blockSize: String): H2OConf = set(PROP_EXTERNAL_COMMUNICATION_BLOCK_SIZE._1, blockSize)
   def setExternalBackendStopTimeout(timeout: Int): H2OConf = set(PROP_EXTERNAL_BACKEND_STOP_TIMEOUT._1, timeout.toString)
-  private[backends] def setExternalBackendFlatFileName(name: String): H2OConf = set(PROP_EXTERNAL_FLATFILE_NAME._1, name)
 
   def externalConfString: String =
     s"""Sparkling Water configuration:
@@ -214,7 +213,4 @@ object ExternalBackendConf {
 
   /** Disable version check of external H2O backend */
   val PROP_EXTERNAL_DISABLE_VERSION_CHECK = ("spark.ext.h2o.external.disable.version.check", false)
-
-  /** Name of file where flatfile is stored in case of automatic mode of external backend */
-  val PROP_EXTERNAL_FLATFILE_NAME = ("spark.ext.h2o.external.flatfile.name", None)
 }
