+++ core/src/main/scala/org/apache/spark/h2o/backends/external/ExternalBackendUtils.scala
@@ -17,47 +17,13 @@
 
 package org.apache.spark.h2o.backends.external
 
-import java.io.File
-
-import org.apache.spark.h2o.H2OConf
-import org.apache.spark.h2o.backends.{ArgumentBuilder, SharedBackendUtils}
-import water.{ExternalFrameUtils, H2O, Paxos}
+import org.apache.spark.h2o.{H2OConf, H2OContext}
+import org.apache.spark.h2o.backends.SharedBackendUtils
+import org.apache.spark.h2o.utils.NodeDesc
+import water.api.RestAPIManager
+import water.{ExternalFrameUtils, H2O, H2OStarter, Paxos}
 
 private[backends] trait ExternalBackendUtils extends SharedBackendUtils {
-
-
-  protected[backends] def waitForCloudSize(expectedSize: Int, timeoutInMilliseconds: Long): Int = {
-    val start = System.currentTimeMillis()
-    while (System.currentTimeMillis() - start < timeoutInMilliseconds) {
-      if (H2O.CLOUD.size() >= expectedSize && Paxos._commonKnowledge) {
-        return H2O.CLOUD.size()
-      }
-      try {
-        Thread.sleep(100)
-      } catch {
-        case _: InterruptedException =>
-      }
-    }
-    H2O.CLOUD.size()
-  }
-
-  /**
-    * Get arguments for H2O client
-    *
-    * @return array of H2O client arguments.
-    */
-  override def getH2OClientArgs(conf: H2OConf): Seq[String] = {
-    val builder = new ArgumentBuilder()
-      .add(super.getH2OClientArgs(conf))
-
-    if (conf.isAutoClusterStartUsed) {
-      builder.add("-flatfile", new File(conf.externalBackendFlatFileName().get).getAbsolutePath)
-    } else {
-      builder.add("-flatfile", conf.h2oCluster.map(clusterStr => saveFlatFileAsFile(clusterStr).getAbsolutePath))
-    }
-    builder.buildArgs()
-
-  }
 
   def prepareExpectedTypes(classes: Array[Class[_]]): Array[Byte] = {
     classes.map { clazz =>
@@ -89,6 +55,54 @@ private[backends] trait ExternalBackendUtils extends SharedBackendUtils {
     }
   }
 
+
+  protected[backends] def startH2OClient(hc: H2OContext, nodes: Array[NodeDesc]): Unit = {
+    val conf = hc.getConf
+    val args = getH2OClientArgs(conf).toArray
+    val launcherArgs = toH2OArgs(args, nodes)
+    logDebug(s"Arguments used for launching the H2O client node: ${launcherArgs.mkString(" ")}")
+
+    H2OStarter.start(launcherArgs, false)
+
+    val expectedSize = conf.clusterSize.get.toInt
+    val discoveredSize = waitForCloudSize(expectedSize, conf.cloudTimeout)
+    if (discoveredSize < expectedSize) {
+      if (conf.isAutoClusterStartUsed) {
+        logError(s"Exiting! External H2O cluster was of size $discoveredSize but expected was $expectedSize!!")
+        hc.stop(stopSparkContext = true)
+      }
+      throw new RuntimeException("Cloud size " + discoveredSize + " under " + expectedSize);
+    }
+
+    RestAPIManager(hc).registerAll()
+    H2O.startServingRestApi()
+  }
+
+  protected[backends] def verifyH2OClientCloudUp(conf: H2OConf, nodes: Array[NodeDesc]): Unit = {
+    val cloudMembers = H2O.CLOUD.members()
+    if (cloudMembers.isEmpty) {
+      if (conf.isManualClusterStartUsed) {
+        throw new H2OClusterNotRunning(
+          s"""
+             |External H2O cluster is not running or could not be connected to. Provided configuration:
+             |  cluster name            : ${conf.cloudName.get}
+             |  cluster representative  : ${conf.h2oCluster.getOrElse("Using multi-cast discovery!")}
+             |  cluster start timeout   : ${conf.clusterStartTimeout} sec
+             |
+             |It is possible that in case you provided only the cluster name, h2o is not able to cloud up
+             |because multi-cast communication is limited in your network. In that case, please consider starting the
+             |external H2O cluster with flatfile and set the following configuration '${ExternalBackendConf.PROP_EXTERNAL_CLUSTER_REPRESENTATIVE._1}'
+        """.stripMargin)
+      } else {
+        throw new H2OClusterNotRunning("Problem with connecting to external H2O cluster started on yarn." +
+          "Please check the YARN logs.")
+      }
+    }
+    if (cloudMembers.size != nodes.length) {
+      throw new RuntimeException("Invalid number of discovered H2O nodes.")
+    }
+  }
+
   protected[backends] def launchShellCommand(cmdToLaunch: Seq[String]): Int = {
     import scala.sys.process._
     val processOut = new StringBuffer()
@@ -108,4 +122,20 @@ private[backends] trait ExternalBackendUtils extends SharedBackendUtils {
     logError(processErr.toString)
     proc
   }
+
+  private def waitForCloudSize(expectedSize: Int, timeoutInMilliseconds: Long): Int = {
+    val start = System.currentTimeMillis()
+    while (System.currentTimeMillis() - start < timeoutInMilliseconds) {
+      if (H2O.CLOUD.size() >= expectedSize && Paxos._commonKnowledge) {
+        return H2O.CLOUD.size()
+      }
+      try {
+        Thread.sleep(100)
+      } catch {
+        case _: InterruptedException =>
+      }
+    }
+    H2O.CLOUD.size()
+  }
+
 }
